container/requirements.txt
requests
beautifulsoup4
python-dotenv
loguru
pydantic
fake-useragent
azure-storage-blob
